+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 * Unreleased
+* v1.0.4 (2022-01-18)
     * Add `EspSntpClock` class which configures the SNTP client on the ESP8266
       and ESP32 platforms, sets the `configTime()` timezone to UTC, and exposes
       the `time()` function through the `Clock` interface.